+++ resources/ts/components/Layout.vue
